Migrate several more fusionpbx apps/mods to granular permissions.
git-svn-id: http://fusionpbx.googlecode.com/svn/trunk/fusionpbx@1099 e2954ef0-caed-11de-bd98-e341d93cc3cc

+++ mod/call_broadcast/v_call_broadcast.php
@@ -26,7 +26,7 @@
 include "root.php";
 require_once "includes/config.php";
 require_once "includes/checkauth.php";
-if (ifgroup("admin") || ifgroup("tenant")) {
+if (permission_exists('call_broadcast_view')) {
 	//access granted
 }
 else {
@@ -94,7 +94,9 @@ require_once "includes/paging.php";
 	echo thorderby('broadcast_desc', 'Description', $orderby, $order);
 	//echo thorderby('recordingid', 'Recording', $orderby, $order);
 	echo "<td align='right' width='42'>\n";
-	echo "	<a href='v_call_broadcast_edit.php' alt='add'>$v_link_label_add</a>\n";
+	if (permission_exists('call_broadcast_add')) {
+		echo "	<a href='v_call_broadcast_edit.php' alt='add'>$v_link_label_add</a>\n";
+	}
 	echo "</td>\n";
 	echo "<tr>\n";
 
@@ -109,10 +111,12 @@ require_once "includes/paging.php";
 			//echo "	<td valign='top' class='".$rowstyle[$c]."'>".$row[recordingid]."</td>\n";
 			echo "	<td valign='top' class='".$rowstyle[$c]."'>".$row[broadcast_desc]."&nbsp;</td>\n";
 			echo "	<td valign='top' align='right'>\n";
-			echo "		<a href='v_call_broadcast_edit.php?id=".$row[call_broadcast_id]."' alt='edit'>$v_link_label_edit</a>\n";
-			echo "		<a href='v_call_broadcast_delete.php?id=".$row[call_broadcast_id]."' alt='delete' onclick=\"return confirm('Do you really want to delete this?')\">$v_link_label_delete</a>\n";
-			//echo "		<input type='button' class='btn' name='' alt='edit' onclick=\"window.location='v_call_broadcast_edit.php?id=".$row[call_broadcast_id]."'\" value='e'>\n";
-			//echo "		<input type='button' class='btn' name='' alt='delete' onclick=\"if (confirm('Are you sure you want to delete this?')) { window.location='v_call_broadcast_delete.php?id=".$row[call_broadcast_id]."' }\" value='x'>\n";
+			if (permission_exists('call_broadcast_edit')) {
+				echo "		<a href='v_call_broadcast_edit.php?id=".$row[call_broadcast_id]."' alt='edit'>$v_link_label_edit</a>\n";
+			}
+			if (permission_exists('call_broadcast_delete')) {
+				echo "		<a href='v_call_broadcast_delete.php?id=".$row[call_broadcast_id]."' alt='delete' onclick=\"return confirm('Do you really want to delete this?')\">$v_link_label_delete</a>\n";
+			}
 			echo "	</td>\n";
 			echo "</tr>\n";
 			if ($c==0) { $c=1; } else { $c=0; }
@@ -128,7 +132,9 @@ require_once "includes/paging.php";
 	echo "		<td width='33.3%' nowrap>&nbsp;</td>\n";
 	echo "		<td width='33.3%' align='center' nowrap>$pagingcontrols</td>\n";
 	echo "		<td width='33.3%' align='right'>\n";
-	echo "	<a href='v_call_broadcast_edit.php' alt='add'>$v_link_label_add</a>\n";
+	if (permission_exists('call_broadcast_add')) {
+		echo "		<a href='v_call_broadcast_edit.php' alt='add'>$v_link_label_add</a>\n";
+	}
 	echo "		</td>\n";
 	echo "	</tr>\n";
  	echo "	</table>\n";
