+++ mod/call_center/v_call_center_agent.php
@@ -26,7 +26,7 @@
 require_once "root.php";
 require_once "includes/config.php";
 require_once "includes/checkauth.php";
-if (ifgroup("admin") || ifgroup("superadmin")) {
+if (permission_exists('call_center_agents_view')) {
 	//access granted
 }
 else {
@@ -121,7 +121,9 @@ require_once "includes/paging.php";
 	//echo thorderby('agent_reject_delay_time', 'Reject Delay Time', $orderby, $order);
 	//echo thorderby('agent_busy_delay_time', 'Busy Delay Time', $orderby, $order);
 	echo "<td align='right' width='42'>\n";
-	echo "	<a href='v_call_center_agent_edit.php' alt='add'>$v_link_label_add</a>\n";
+	if (permission_exists('call_center_agents_add')) {
+		echo "	<a href='v_call_center_agent_edit.php' alt='add'>$v_link_label_add</a>\n";
+	}
 	echo "</td>\n";
 	echo "<tr>\n";
 
@@ -141,8 +143,12 @@ require_once "includes/paging.php";
 			//echo "	<td valign='top' class='".$row_style[$c]."'>".$row[agent_reject_delay_time]."&nbsp;</td>\n";
 			//echo "	<td valign='top' class='".$row_style[$c]."'>".$row[agent_busy_delay_time]."&nbsp;</td>\n";
 			echo "	<td valign='top' align='right'>\n";
-			echo "		<a href='v_call_center_agent_edit.php?id=".$row[call_center_agent_id]."' alt='edit'>$v_link_label_edit</a>\n";
-			echo "		<a href='v_call_center_agent_delete.php?id=".$row[call_center_agent_id]."' alt='delete' onclick=\"return confirm('Do you really want to delete this?')\">$v_link_label_delete</a>\n";
+			if (permission_exists('call_center_agents_edit')) {
+				echo "		<a href='v_call_center_agent_edit.php?id=".$row[call_center_agent_id]."' alt='edit'>$v_link_label_edit</a>\n";
+			}
+			if (permission_exists('call_center_agents_delete')) {
+				echo "		<a href='v_call_center_agent_delete.php?id=".$row[call_center_agent_id]."' alt='delete' onclick=\"return confirm('Do you really want to delete this?')\">$v_link_label_delete</a>\n";
+			}
 			//echo "		<input type='button' class='btn' name='' alt='edit' onclick=\"window.location='v_call_center_agent_edit.php?id=".$row[call_center_agent_id]."'\" value='e'>\n";
 			//echo "		<input type='button' class='btn' name='' alt='delete' onclick=\"if (confirm('Are you sure you want to delete this?')) { window.location='v_call_center_agent_delete.php?id=".$row[call_center_agent_id]."' }\" value='x'>\n";
 			echo "	</td>\n";
@@ -159,7 +165,9 @@ require_once "includes/paging.php";
 	echo "		<td width='33.3%' nowrap>&nbsp;</td>\n";
 	echo "		<td width='33.3%' align='center' nowrap>$paging_controls</td>\n";
 	echo "		<td width='33.3%' align='right'>\n";
-	echo "			<a href='v_call_center_agent_edit.php' alt='add'>$v_link_label_add</a>\n";
+	if (permission_exists('call_center_agents_add')) {
+		echo "			<a href='v_call_center_agent_edit.php' alt='add'>$v_link_label_add</a>\n";
+	}
 	echo "		</td>\n";
 	echo "	</tr>\n";
  	echo "	</table>\n";
