+++ mod/call_center/v_call_center_queue.php
@@ -26,7 +26,7 @@
 require_once "root.php";
 require_once "includes/config.php";
 require_once "includes/checkauth.php";
-if (ifgroup("admin") || ifgroup("superadmin")) {
+if (permission_exists('call_center_queues_view')) {
 	//access granted
 }
 else {
@@ -107,7 +107,6 @@ require_once "includes/paging.php";
 
 	echo "<div align='center'>\n";
 	echo "<table width='100%' border='0' cellpadding='0' cellspacing='0'>\n";
-
 	echo "<tr>\n";
 	echo thorderby('queue_name', 'Queue Name', $orderby, $order);
 	echo thorderby('queue_extension', 'Extension', $orderby, $order);
@@ -125,7 +124,9 @@ require_once "includes/paging.php";
 	//echo thorderby('queue_tier_rule_wait_multiply_level', 'Tier Rule Wait Multiply Level', $orderby, $order);
 	echo thorderby('queue_description', 'Description', $orderby, $order);
 	echo "<td align='right' width='42'>\n";
-	echo "	<a href='v_call_center_queue_edit.php' alt='add'>$v_link_label_add</a>\n";
+	if (permission_exists('call_center_queues_add')) {
+		echo "	<a href='v_call_center_queue_edit.php' alt='add'>$v_link_label_add</a>\n";
+	}
 	echo "</td>\n";
 	echo "<tr>\n";
 
@@ -148,10 +149,12 @@ require_once "includes/paging.php";
 			//echo "	<td valign='top' class='".$row_style[$c]."'>".$row[queue_tier_rule_wait_multiply_level]."&nbsp;</td>\n";
 			echo "	<td valign='top' class='".$row_style[$c]."'>".$row[queue_description]."&nbsp;</td>\n";
 			echo "	<td valign='top' align='right'>\n";
-			echo "		<a href='v_call_center_queue_edit.php?id=".$row[call_center_queue_id]."' alt='edit'>$v_link_label_edit</a>\n";
-			echo "		<a href='v_call_center_queue_delete.php?id=".$row[call_center_queue_id]."' alt='delete' onclick=\"return confirm('Do you really want to delete this?')\">$v_link_label_delete</a>\n";
-			//echo "		<input type='button' class='btn' name='' alt='edit' onclick=\"window.location='v_call_center_queue_edit.php?id=".$row[call_center_queue_id]."'\" value='e'>\n";
-			//echo "		<input type='button' class='btn' name='' alt='delete' onclick=\"if (confirm('Are you sure you want to delete this?')) { window.location='v_call_center_queue_delete.php?id=".$row[call_center_queue_id]."' }\" value='x'>\n";
+			if (permission_exists('call_center_queues_edit')) {
+				echo "		<a href='v_call_center_queue_edit.php?id=".$row[call_center_queue_id]."' alt='edit'>$v_link_label_edit</a>\n";
+			}
+			if (permission_exists('call_center_queues_delete')) {
+				echo "		<a href='v_call_center_queue_delete.php?id=".$row[call_center_queue_id]."' alt='delete' onclick=\"return confirm('Do you really want to delete this?')\">$v_link_label_delete</a>\n";
+			}
 			echo "	</td>\n";
 			echo "</tr>\n";
 			if ($c==0) { $c=1; } else { $c=0; }
@@ -166,7 +169,9 @@ require_once "includes/paging.php";
 	echo "		<td width='33.3%' nowrap>&nbsp;</td>\n";
 	echo "		<td width='33.3%' align='center' nowrap>$paging_controls</td>\n";
 	echo "		<td width='33.3%' align='right'>\n";
-	echo "			<a href='v_call_center_queue_edit.php' alt='add'>$v_link_label_add</a>\n";
+	if (permission_exists('call_center_queues_add')) {
+		echo "			<a href='v_call_center_queue_edit.php' alt='add'>$v_link_label_add</a>\n";
+	}
 	echo "		</td>\n";
 	echo "	</tr>\n";
  	echo "	</table>\n";
